+++ src/App.js
@@ -1,0 +1,18 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import Header from "./components/Header";
+import Body from "./components/Body";
+import Footer from "./components/Footer";
+
+const AppLayout = () => {
+  return (
+    <div>
+      <Header />
+      <Body />
+      <Footer year={2025} companyName="Online Food Delivery" />
+    </div>
+  );
+};
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(<AppLayout />);
