+++ src/components/Body.js
@@ -1,0 +1,16 @@
+import RestaurentCard from "./RestaurentCard";
+import resList from "../utils/mockData";
+const Body = () => {
+  return (
+    <div className="body">
+      <div className="search">Search</div>
+      <div className="res-container">
+        {resList.map(restaurant =>
+          <RestaurentCard key={restaurant.info.id} resData={restaurant} />
+        )}
+      </div>
+    </div>
+  );
+};
+
+export default Body;
