+++ src/components/Header.js
@@ -1,0 +1,19 @@
+import { LOGO_URL } from "../utils/constants"; //since we are importing named export from constants.js we have to use curly braces
+const Header = () => {
+  return (
+    <div className="header">
+      <div className="logo">
+        <img src={LOGO_URL} alt="Logo" />
+      </div>
+      <div className="nav-items">
+        <ul>
+          <li>Home</li>
+          <li>About</li>
+          <li>Contact</li>
+        </ul>
+      </div>
+    </div>
+  );
+};
+
+export default Header;
