+++ src/utils/constants.js
@@ -1,0 +1,7 @@
+// Constants for the application it is exported as a named export
+
+export const LOGO_URL =
+  "https://thumbs.dreamstime.com/b/food-delivery-logo-template-vector-icon-illustration-170869600.jpg";
+
+export const SWIGGY_CDN_URL =
+  "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/";
