+++ src/utils/mockData.js
@@ -1,26 +1,3 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-
-const Header = () => {
-  return (
-    <div className="header">
-      <div className="logo">
-        <img
-          src="https://thumbs.dreamstime.com/b/food-delivery-logo-template-vector-icon-illustration-170869600.jpg"
-          alt="Logo"
-        />
-      </div>
-      <div className="nav-items">
-        <ul>
-          <li>Home</li>
-          <li>About</li>
-          <li>Contact</li>
-        </ul>
-      </div>
-    </div>
-  );
-};
-
 const resList = [
   {
     info: {
@@ -772,94 +749,4 @@ const resList = [
     }
   }
 ];
-
-console.log(resList);
-const RestaurentCard = props => {
-  const { resData } = props;
-  const {
-    cloudinaryImageId,
-    name,
-    cuisines,
-    avgRating,
-    costForTwo,
-    sla,
-    locality,
-    areaName
-  } = resData.info;
-  return (
-    <div className="res-card">
-      <img
-        className="res-image"
-        src={
-          "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/" +
-          cloudinaryImageId
-        } // Using template literals to construct the image URL"
-        alt="Restaurant"
-      />
-      <h3 className="res-name">
-        {name}
-      </h3>
-      <h4>
-        <span className="res-cuisine">
-          Cuisine: {cuisines.join(", ")}
-        </span>
-      </h4>
-      <span className="res-rating">
-        Rating : {avgRating}
-      </span>
-      <h4 />
-      <h4>
-        <span className="res-cost">
-          Cost per 2 person: {costForTwo}
-        </span>
-      </h4>
-      <h4>
-        <span className="res-delivery-time">
-          Delivery Time: {sla.deliveryTime} mins
-        </span>
-      </h4>
-      <h4>
-        <span className="res-locality">
-          Locality: {locality}, {areaName}
-        </span>
-      </h4>
-    </div>
-  );
-};
-
-const Body = () => {
-  return (
-    <div className="body">
-      <div className="search">Search</div>
-      <div className="res-container">
-        {resList.map(restaurant =>
-          <RestaurentCard key={restaurant.info.id} resData={restaurant} />
-        )}
-      </div>
-    </div>
-  );
-};
-
-const Footer = props => {
-  const { year, companyName } = props;
-  return (
-    <div className="footer">
-      <p className="footer-text">
-        © {year} {companyName}
-      </p>
-    </div>
-  );
-};
-
-const AppLayout = () => {
-  return (
-    <div>
-      <Header />
-      <Body />
-      <Footer year={2025} companyName="Online Food Delivery" />
-    </div>
-  );
-};
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<AppLayout />);
+export default resList;
